BFPD/submission-pemda/requirements.txt
sqlalchemy~=2.0
psycopg2-binary~=2.9
pandas~=2.2
numpy~=1.26
requests~=2.32
beautifulsoup4~=4.12
google-auth~=2.36
google-api-python-client~=2.152
pytest~=8.1
pytest-cov~=4.1
pytest-mock~=3.12
python-dotenv~=1.0
python-dateutil~=2.9